Restructure background and add voting content
As I write the voting systems section it's become clear that I need some
background content on forms of governance first. I was going to add a
separate section on governance, but it interrupts the flow too much to
have a governance and voting chapter. Instead I'll move the governance,
voting, and blockchain sections into one background section. This is
probably how it should be anyways. We'll see how moving the blockchain
section before the literature section will influence the literature
review.